Longest subarray without repeating characters
Approach: We start traversing the string from left to right and maintain track of:
the current substring with non-repeating characters with the help of a start and end index
the longest non-repeating substring output
a lookup table of already visited characters
For every new character, we look for it in the already visited characters. If the character has already been visited and is part of the current substring with non-repeating characters, we update the start index. Otherwise, we'll continue traversing the string.

Since we are traversing the string only once, the time complexity will be linear, or O(n).
This approach is also known as the sliding window pattern.