Find intersection point of Y linked list
Approach: It can be observed that the number of nodes in traversing the first linked list and then from the head of the second linked list to intersection point is equal to the number of nodes involved in traversing the second linked list and then from head of the first list to the intersection point. Considering the example given above, start traversing the two linked lists with two pointers curr1 and curr2 pointing to the heads of the given linked lists respectively.

1)If curr1 != null then update it to point to the next node, else it is updated to point to the first node of the second list.
2)If curr2 != null then update it to point to the next node, else it is updated to point to the first node of the first list.
3)Repeat the above steps while curr1 is not equal to curr2.

The two pointers curr1 and curr2 will be pointing to the same node now i.e. the merging point.